backend/app/auth/jwt_handler.py
# backend/app/auth/jwt_handler.py
import jwt
from datetime import datetime, timedelta
from app.config import JWT_SECRET, JWT_ALGORITHM

def create_access_token(subject: str, expires_minutes: int = 60*24):
    payload = {
        "sub": subject,
        "exp": datetime.now() + timedelta(minutes=expires_minutes)
    }
    return jwt.encode(payload, JWT_SECRET, algorithm=JWT_ALGORITHM)

def decode_token(token: str):
    try:
        payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        return payload.get("sub")
    except Exception:
        return None
