backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
passlib[bcrypt]
pyjwt
aiokafka
python-dotenv